backend/workers/requirements.txt
requests>=2.31.0
youtube-transcript-api>=0.6.2
psycopg2-binary>=2.9.9

